Improved fix for flickering using a setTimeout, refactored into util class
