+++ src/components/SketchesNav/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
+import DelayedSort from '../../utils/DelayedSort'
 import PropTypes from 'prop-types'
 import NavItem from '../NavItem'
 import styled from 'styled-components'
 import { SortableContainer, SortableElement } from 'react-sortable-hoc'
-import arrayMove from 'array-move'
 
 const Wrapper = styled.nav`
   margin-bottom: 2rem;
@@ -33,38 +33,10 @@ const SortableList = SortableContainer(({ items, ...props }) =>
   </ul>
 )
 
-// This sketch can't be stateless because of the reordering
-// Reordering works stateless, but with some annoying flickering
-// It's quite a bit of work just to stop a tiny little thing, but it was annoying!
-// https://github.com/clauderic/react-sortable-hoc/issues/280
-class SketchesNav extends React.Component {
-  state = {
-    items: this.props.items,
-    sceneId: this.props.sceneId,
-  }
-
-  // We update both internal and external state onSortEnd
-  handleOnSortEnd = ({ oldIndex, newIndex }) => {
-    this.setState(({ items }) => ({
-      items: arrayMove(items, oldIndex, newIndex),
-    }))
-    // This wont affect the rendering of the component but we need to store the state properly in redux!
+// Using DelayedSort helper util to stop flickering
+class SketchesNav extends DelayedSort {
+  updateExternalStateAfterSort (oldIndex, newIndex) {
     this.props.onSortEnd(this.props.sceneId, oldIndex, newIndex)
-  }
-
-  // We update the state based on props if the scene has changed or an item has been added/deleted
-  // If only the order has changed, it will just use the internal state (stops the flickering)
-  static getDerivedStateFromProps (props, state) {
-    if (
-      (props.sceneId !== state.sceneId) || // has the scene changed?
-      (props.items && state.items && props.items.length !== state.items.length) // has an item been added or removed?
-    ) {
-      return {
-        items: props.items,
-        sceneId: props.sceneId,
-      }
-    }
-    return null
   }
 
   render () {
